// frontend/src/app/components/interfaces/Usuario.ts
export interface Usuario {
    primerApellido: string;
    segundoApellido: string;
    primerNombre: string;
    segundoNombre: string;
    pais: string;
    tipoIdentificacion: string;
    numeroIdentificacion: string;
    fechaIngreso: string;
    estado: string;
    correo: string;
    area: string;
  }